Projects/QRcode_Barcode_Generator/requirements.txt
Django>=4.0,<5.0
pandas>=2.0,<3.0
openpyxl>=3.0,<4.0
barcode>=1.1.0,<2.0
qrcode>=7.3.1,<8.0
Pillow>=9.0,<10.0
